requirements.txt
streamlit==1.51.0
pandas==2.3.3
numpy==2.3.5
plotly==6.4.0
kafka-python==2.2.15
pymongo==4.6.1
requests==2.32.3
streamlit-autorefresh==1.0.1
